change from integer to bigdecimal to manage amount value

+++ src/main/java/it/gov/pagopa/rtp/activator/domain/rtp/Rtp.java
@@ -1,9 +1,10 @@
 package it.gov.pagopa.rtp.activator.domain.rtp;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-public record Rtp(String noticeNumber, Integer amount, String description, LocalDate expiryDate, String payerId,
+public record Rtp(String noticeNumber, BigDecimal amount, String description, LocalDate expiryDate, String payerId,
         String payeeName, String payeeId,
         ResourceID resourceID, LocalDateTime savingDateTime,
         String rtpSpId, String endToEndId, String iban, String payTrxRef, String flgConf) {
